Progress on sprite animation functionality

+++ enginev2/graphics/animation/Sprite.cpp
@@ -1,6 +1,6 @@
 #include "Sprite.h"
 
-Sprite::Sprite(std::shared_ptr<Texture> texture, const std::vector<Rectangle>& frames, float animationTime):
+Sprite::Sprite(std::shared_ptr<Texture> texture, const std::vector<Rectangle>& frames, uint64_t animationTime):
 	texturePtr(texture),
 	frames(frames),
 	currentFrame(0),
